refactor: promotion 인터페이스들과 그들의 구현 클래스의 패키징을 수정
