+++ src/main/java/christmas/domain/promotion/context/discount/DiscountPromotion.java
@@ -1,6 +1,7 @@
-package christmas.domain.promotion.precondition;
+package christmas.domain.promotion.context.discount;
 
 import christmas.domain.model.classes.decemberEventPlan.DecemberEventPlan;
+import christmas.domain.promotion.precondition.ChristmasPromotionPrecondition;
 import christmas.dto.DiscountInfo;
 
 import java.util.Optional;
