+++ src/main/java/christmas/domain/promotion/context/discount/impl/WeekdayPromotion.java
@@ -1,9 +1,9 @@
-package christmas.domain.promotion.context.discount;
+package christmas.domain.promotion.context.discount.impl;
 
 import christmas.domain.model.classes.decemberEventPlan.DecemberEventPlan;
+import christmas.domain.promotion.context.discount.DiscountPromotion;
 import christmas.domain.promotion.enums.PromotionName;
 import christmas.domain.promotion.precondition.ChristmasPromotionPrecondition;
-import christmas.domain.promotion.precondition.DiscountPromotion;
 import christmas.domain.promotion.strategy.dateCheckStrategy.impl.DayOfWeekCondition;
 import christmas.domain.promotion.strategy.discountStrategy.byDecemberEventPlan.DiscountByDecemberEventPlan;
 import christmas.dto.DiscountInfo;
