+++ src/main/java/christmas/domain/promotion/context/giveaway/GiveawayPromotion.java
@@ -1,6 +1,7 @@
-package christmas.domain.promotion.precondition;
+package christmas.domain.promotion.context.giveaway;
 
 import christmas.domain.model.classes.decemberEventPlan.DecemberEventPlan;
+import christmas.domain.promotion.precondition.ChristmasPromotionPrecondition;
 import christmas.dto.GiveawayAndQuantityDto;
 
 import java.util.Optional;
