+++ src/main/java/christmas/domain/promotion/strategy/discountStrategy/byDecemberEventPlan/DiscountByDecemberEventPlan.java
@@ -3,8 +3,8 @@ package christmas.domain.promotion.strategy.discountStrategy.byDecemberEventPlan
 import christmas.domain.model.classes.decemberEventPlan.DecemberEventPlan;
 import christmas.domain.promotion.strategy.discountStrategy.DiscountStrategy;
 
-import static christmas.domain.promotion.context.discount.WeekdayPromotion.WEEKDAY_DISCOUNT_AMOUNT;
-import static christmas.domain.promotion.context.discount.WeekendPromotion.WEEKEND_DISCOUNT_AMOUNT;
+import static christmas.domain.promotion.context.discount.impl.WeekdayPromotion.WEEKDAY_DISCOUNT_AMOUNT;
+import static christmas.domain.promotion.context.discount.impl.WeekendPromotion.WEEKEND_DISCOUNT_AMOUNT;
 
 public enum DiscountByDecemberEventPlan implements DiscountStrategy {
     BY_DESSERT_COUNT(decemberEventPlan -> WEEKDAY_DISCOUNT_AMOUNT * decemberEventPlan.countDessertMenus()),
